rename config-injectors to config-tokens
