+++ projects/core/src/anonymous-consents/config/anonymous-consents-config.ts
@@ -1,6 +1,6 @@
 import { OccConfig } from '../../occ/config/occ-config';
 import { Injectable } from '@angular/core';
-import { Config } from '../../config/config-injectors';
+import { Config } from '../../config/config-tokens';
 
 @Injectable({
   providedIn: 'root',
