+++ projects/core/src/config/config-tokens.spec.ts
@@ -7,7 +7,7 @@ import {
   RootConfig,
 } from '@spartacus/core';
 
-describe('Config Injectors', () => {
+describe('Config Injector Tokens', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       providers: [
