+++ projects/core/src/features-config/config/features-config.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Config } from '../../config/config-injectors';
+import { Config } from '../../config/config-tokens';
 
 @Injectable({
   providedIn: 'root',
