+++ projects/core/src/global-message/config/global-message-config.ts
@@ -1,6 +1,6 @@
 import { GlobalMessageType } from '../models/global-message.model';
 import { Injectable } from '@angular/core';
-import { Config } from '../../config/config-injectors';
+import { Config } from '../../config/config-tokens';
 
 export type GlobalMessageTypeConfig = {
   timeout?: number;
