+++ projects/core/src/i18n/config/i18n-config.ts
@@ -1,6 +1,6 @@
 import { TranslationResources } from '../translation-resources';
 import { Injectable } from '@angular/core';
-import { Config } from '../../config/config-injectors';
+import { Config } from '../../config/config-tokens';
 
 @Injectable({
   providedIn: 'root',
