+++ projects/core/src/occ/config/occ-config.ts
@@ -2,7 +2,7 @@ import { SiteContextConfig } from '../../site-context/config/site-context-config
 import { OccEndpoints } from '../occ-models/occ-endpoints.model';
 import { LoadingScopes } from './loading-scopes-config';
 import { Injectable } from '@angular/core';
-import { Config } from '../../config/config-injectors';
+import { Config } from '../../config/config-tokens';
 
 @Injectable({
   providedIn: 'root',
