+++ projects/core/src/routing/configurable-routes/config/routing-config.ts
@@ -1,6 +1,6 @@
 import { RoutesConfig } from '../routes-config';
 import { Injectable } from '@angular/core';
-import { Config } from '../../../config/config-injectors';
+import { Config } from '../../../config/config-tokens';
 
 export const enum RouteLoadStrategy {
   /**
