+++ projects/core/src/store-finder/config/store-finder-config.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Config } from '../../config/config-injectors';
+import { Config } from '../../config/config-tokens';
 
 @Injectable({
   providedIn: 'root',
